// src/store/modules/songs/mutation-types.ts
export enum SongsMutationTypes {
    LANGUAGE = "SONGS_LANGUAGE",
    TRANSCODE = "SONGS_TRANSCODE",
    COLLECTIONS = "SONGS_COLLECTIONS",
    COLLECTION = "SONGS_COLLECTION",
    
    CONTRIBUTOR = "SONGS_CONTRIBUTOR",

    SET_LIST = "SONGS_LIST",
    SET_SONG_NUMBER = "SONGS_SET_SONG_NUMBER",
    SET_SONG = "SONGS_SET_SONG",
    SET_SONGS = "SONGS_SET_SONGS",

    // SET_LYRICS = "SONGS_SET_LYRICS",
    //SET_LYRICS_TRANSPOSED = "SONGS_SET_LYRICS_TRANSPOSED",

    SET_VERSES = "SONGS_VERSES",
    SET_LINES = "SONGS_LINES",
    SET_TRANSPOSITION = "SONGS_TRANSPOSITION",

    SET_VIEW = "SONGS_SET_VIEW",
    SET_FILTER = "SONGS_SET_FILTER",

    SET_AUDIO = "SONGS_SET_AUDIO",

    SET_SHEETMUSIC_TRANSPOSITION = "SONGS_SHEETMUSIC_TRANSPOSITION",
    SET_SHEETMUSIC_OPTIONS = "SONGS_SET_SHEETMUSIC_OPTIONS",

    SEARCH = "SONGS_SEARCH",
    SEARCH_RESULT = "SONGS_SEARCH_RESULT",
}